Implemented initial chat message architecture plus refactors and a potential beginning for mail/letters.
Added a **ChatMessageStore.cs** object which will hold the last 10 messages sent in chat, this is 10 because the game source code sets the cap as 10.
This object uses the singleton pattern to be able to ensure we only ever have one static list of messages available at any time.
It will compare the stored messages against the newly received messages and only update if it identifies a change.

When it detects new messages, it adds them to a queue set up in the new *ChatMessageHandler.cs** class.
This class is a glorified queue which enqueues, dequeues, processes next and processes all.
Currently nothing is happening except a text message to the screen for debugging purposes.

The intention is to use a manager or factory style pattern to be able to handle commands appropriately with separation of concerns.

We are updating the message store every one second tick providing the game world is ready (therefore chat box is ready).
We only check to process the messages if server mode / toggle is turned on - this means admins can prevent message spam by turning the server mode off.

It also means commands are only executed at one second intervals (by placement in one second update tick)

TODO: An implementation to be able to cap messages or identify spam, in order to prevent crashes or loops.

**Added some mail / letter logic and enums**
This I am not too sure about, I'll revisit it... I need to test more of the events and letters in game with a debugger to think about the approaches I can use.

**Refactoring**
This is mainly just moving files around and thinking of the naming of the directories.